--- src/screen/denuncia/Denuncia.tsx
@@ -1,7 +1,0 @@
-import DenunciaModal from "../../components/DenunciaModal";
-
-export default function Denuncia(){
-    return(
-        <DenunciaModal modalVisivel={true}/>
-    )
-}
